I made changes to User, Manager, Staff class
These changes makes sure that we remember user information such as
first and last name, and id. So that we can use this information within
our query.

So the To-Do list is:
Add functionality to all button for User, Manager, and Staff panel.
Modify some queries to support minimum requirements, which are Outer
Join, Correlated Subquery, and a Set Operation.  We have met other
requirements. But we can always double check before we submit.
We have start on the report as we finish the part of this project.